Seeds and factories, nodemailer, auth flow with JWT, serve static files, upload images

+++ src/modules/users/users.service.ts
@@ -1,25 +1,36 @@
-import { Injectable } from '@nestjs/common';
-import { DeleteResult, FindConditions } from 'typeorm';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { DeepPartial, DeleteResult, FindConditions } from 'typeorm';
 import { CommandBus } from '@nestjs/cqrs';
-import { plainToClass } from 'class-transformer';
+import { plainToInstance } from 'class-transformer';
 
 import { UserEntity } from './entities/user.entity';
-import { UpdateUserDto } from './dtos/update-user.dto';
-import { EmailTakenException, UserNotFoundException } from '../../exceptions';
-import { FindAllUserDto } from './dtos/find-all-user.dto';
-import { CreateSettingDto } from './dtos/create-setting.dto';
+import {
+  EmailTakenException,
+  InvalidCredentialsException,
+  UserNotFoundException,
+} from '../../common/exceptions';
 import { UsersRepository } from './users.repository';
 import { CreateSettingCommand } from './handlers/create-setting.handler';
 import { UserSettingsEntity } from './entities/user-settings.entity';
-import { UserRegisterDto } from '../auth/dtos/user-register.dto';
 import { selfGuard } from '../auth/guards';
-import { ContextProvider } from '../../providers';
+import { UsersSettingsRepository } from './users-settings.repository';
+import { Languages } from '../../common/constants';
+import { UserImageEntity } from './entities/user-image.entity';
+import { UsersImageRepository } from './users-image.repository';
+import { S3Service } from '../shared/services/s3.service';
+import { CreateSettingDto, FindAllUserDto } from './dtos';
+import { Crypto } from '../../common/utilities';
+import { RegisterDto } from '../auth/dtos';
+import { ChangePasswordDto } from './dtos/requests/change-password.dto';
 
 @Injectable()
 export class UsersService {
   constructor(
-    private usersRepository: UsersRepository,
-    private commandBus: CommandBus,
+    private readonly usersRepository: UsersRepository,
+    private readonly usersSettingsRepository: UsersSettingsRepository,
+    private readonly usersImageRepository: UsersImageRepository,
+    private readonly s3Service: S3Service,
+    private readonly commandBus: CommandBus,
   ) {}
 
   async totalRepositoryItems(): Promise<number> {
@@ -29,7 +40,8 @@ export class UsersService {
   findAll(filters: FindAllUserDto): Promise<UserEntity[]> {
     const query = this.usersRepository
       .createQueryBuilder('user')
-      .leftJoinAndSelect('user.settings', 'user_settings');
+      .leftJoinAndSelect('user.settings', 'user_settings')
+      .leftJoinAndSelect('user.image', 'user_image');
 
     query.where('user.softDelete = false');
 
@@ -56,31 +68,29 @@ export class UsersService {
     return query.getMany();
   }
 
-  async findOne(findData: FindConditions<UserEntity>): Promise<UserEntity> {
-    const user: UserEntity | undefined = await this.usersRepository.findOne(
-      findData,
-      { relations: ['settings'] },
-    );
-    if (!user) throw new UserNotFoundException();
-
-    return user;
+  async findOne(
+    findData: FindConditions<UserEntity>,
+  ): Promise<UserEntity | undefined> {
+    return await this.usersRepository.findOne(findData, {
+      relations: ['settings', 'image'],
+    });
   }
 
-  async create(createUserDto: UserRegisterDto): Promise<UserEntity> {
+  async create(dto: RegisterDto): Promise<UserEntity> {
     const takenEmail: UserEntity | undefined =
       await this.usersRepository.findOne({
-        email: createUserDto.email,
+        email: dto.email,
       });
-    if (takenEmail) throw new EmailTakenException(createUserDto.email);
+    if (takenEmail) throw new EmailTakenException(dto.email);
 
-    const user: UserEntity = this.usersRepository.create(createUserDto);
+    const user: UserEntity = this.usersRepository.create(dto);
 
     await this.usersRepository.save(user);
 
     user.settings = await this.createSettings(
       user.id,
-      plainToClass(CreateSettingDto, {
-        isEmailVerified: false,
+      plainToInstance(CreateSettingDto, {
+        language: Languages.en,
       }),
     );
 
@@ -88,20 +98,46 @@ export class UsersService {
   }
 
   async update(
-    uuid: string,
-    updateUserDto: UpdateUserDto,
+    user: UserEntity,
+    updateData: DeepPartial<UserEntity>,
   ): Promise<UserEntity> {
-    const user: UserEntity = await this.findOne({ id: uuid });
-
-    selfGuard(ContextProvider.getAuthUser(), user.id);
-
-    return this.usersRepository.save(
-      this.usersRepository.merge(user, updateUserDto),
+    return await this.usersRepository.save(
+      this.usersRepository.merge(user, updateData),
     );
   }
 
+  async updateUserWithGuard(
+    uuid: string,
+    dto: Partial<UserEntity>,
+  ): Promise<UserEntity> {
+    const user: UserEntity | undefined = await this.findOne({ id: uuid });
+    if (!user) throw new UserNotFoundException();
+
+    selfGuard(user.id);
+
+    return this.update(user, dto);
+  }
+
+  async changePassword(
+    user: UserEntity,
+    changePasswordDto: ChangePasswordDto,
+  ): Promise<void> {
+    if (
+      !Crypto.validateHash(user.password, changePasswordDto.currentPassword)
+    ) {
+      throw new InvalidCredentialsException();
+    }
+
+    selfGuard(user.id);
+
+    await this.update(user, {
+      password: Crypto.generateHash(changePasswordDto.newPassword),
+    });
+  }
+
   async remove(uuid: string): Promise<DeleteResult> {
-    await this.findOne({ id: uuid });
+    const user: UserEntity | undefined = await this.findOne({ id: uuid });
+    if (!user) throw new UserNotFoundException();
 
     return this.usersRepository.delete({ id: uuid });
   }
@@ -113,5 +149,65 @@ export class UsersService {
     return this.commandBus.execute<CreateSettingCommand, UserSettingsEntity>(
       new CreateSettingCommand(userId, createSettingDto),
     );
+  }
+
+  async setHashedRt(user: UserEntity, hashedRt: string): Promise<void> {
+    await this.usersRepository.save(
+      this.usersRepository.merge(user, { hashedRt }),
+    );
+  }
+
+  async revokeHashedRt(user: UserEntity): Promise<void> {
+    await this.usersRepository.save(
+      this.usersRepository.merge(user, { hashedRt: null }),
+    );
+  }
+
+  async createImage(
+    user: UserEntity,
+    file: Express.Multer.File,
+  ): Promise<void> {
+    if (user.image) {
+      await this.deleteImage(user.id, user.image);
+      await this.s3Service.deleteFile({
+        moduleName: 'users',
+        proprietaryId: user.id,
+        fileId: user.image.id,
+        fileExtension: user.image.extension,
+      });
+    }
+
+    // save main entity
+    const extension: string = this.s3Service.getFileExtension(
+      file.originalname,
+    );
+    const entity = this.usersImageRepository.create({
+      extension,
+      userId: user.id,
+      user,
+    });
+    const image: UserImageEntity = await this.usersImageRepository.save(entity);
+
+    // upload image to S3
+    await this.s3Service.uploadSingleFile({
+      moduleName: 'users',
+      proprietaryId: user.id,
+      fileId: image.id,
+      file: file,
+    });
+  }
+
+  async deleteImage(
+    userId: string,
+    image: UserImageEntity | null,
+  ): Promise<void> {
+    if (!image) throw new NotFoundException('User has no assigned image');
+    await this.usersImageRepository.delete({ id: image.id });
+    await this.s3Service.deleteFile({
+      moduleName: 'users',
+      proprietaryId: userId,
+      fileId: image.id,
+      fileExtension: image.extension,
+    });
   }
 }
